[dhctl] chore(dhctl-for-commander): expose dhctl commands in pkg/operations package and subpackages
* Moved functions passed to the kingpin run-functions from cmd/dhctl/commands/* to pkg/operations/* packages.
    * What run-functions has been moved:
        * from cmd/dhctl/commands/bootstrap/bootstrap.go to pkg/operations/bootstrap ClusterBootstrapper object.
        * from cmd/dhctl/commands/bootstrap/phase.go to pkg/operations/bootstrap ClusterBootstrapper object.
        * from cmd/dhctl/commands/converge.go to pkg/operations/converge Converger object.

* **Important**: do not eliminate app.* params at this point yet. Instead override app.* using pkg/operations/* Params if Params has been set. This override localized in pkg/operations package and occurs implicitly. This is needed to provide compatible behaviour with current way of app.* usage, until app.* could be completely eliminated.

* Eliminated InitClusterDestroyer, Destroyer is now separately initialized either in the cmd/dhctl/commands/destroy.go or in the pkg/operations/converge Converger object (destroyer needed here to implement abort phase).

* pkg/operations/bootstrap.go which contains functions which are used as steps during bootstrap process has been moved to the pkg/operations/bootstrap/steps.go (with tests steps_test.go).
    * This is needed to prevent circular dependencies between pkg/operations and  pkg/operations/bootstrap.
    * Also this is one of the TODO tasks which were in the pkg/operations/bootstrap/steps.go.

* Moved pkg/operations/auto-converger.go  into the pkg/operations/converge package.

* Moved ConnectToKubernetesAPI helper to the pkg/operations/kube.go (to avoid circular dependencies, as this common helper used in different pkg/operations/* subpackages).

* Created alternative constructor for ssh-client with explicit parameters instead of app.* cli params.

* Added ability to reinitialize singleton ssh-agent instance.
