+++ dhctl/pkg/operations/bootstrap/cluster-bootstrapper-base-infra.go
@@ -1,0 +1,73 @@
+// Copyright 2023 Flant JSC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package bootstrap
+
+import (
+	"fmt"
+
+	"github.com/deckhouse/deckhouse/dhctl/pkg/app"
+	"github.com/deckhouse/deckhouse/dhctl/pkg/config"
+	"github.com/deckhouse/deckhouse/dhctl/pkg/log"
+	"github.com/deckhouse/deckhouse/dhctl/pkg/state"
+	"github.com/deckhouse/deckhouse/dhctl/pkg/state/cache"
+	"github.com/deckhouse/deckhouse/dhctl/pkg/terraform"
+	"github.com/deckhouse/deckhouse/dhctl/pkg/util/tomb"
+)
+
+func (b *ClusterBootstrapper) BaseInfrastructure() error {
+	if restore, err := b.applyParams(); err != nil {
+		return err
+	} else {
+		defer restore()
+	}
+
+	metaConfig, err := config.ParseConfig(app.ConfigPath)
+	if err != nil {
+		return err
+	}
+
+	if metaConfig.ClusterType != config.CloudClusterType {
+		return fmt.Errorf(bootstrapPhaseBaseInfraNonCloudMessage)
+	}
+
+	cachePath := metaConfig.CachePath()
+	if err = cache.Init(cachePath); err != nil {
+		// TODO: it's better to ask for confirmation here
+		return fmt.Errorf(cacheMessage, cachePath, err)
+	}
+
+	stateCache := cache.Global()
+
+	if app.DropCache {
+		stateCache.Clean()
+		stateCache.Delete(state.TombstoneKey)
+	}
+
+	clusterUUID, err := generateClusterUUID(stateCache)
+	if err != nil {
+		return err
+	}
+	metaConfig.UUID = clusterUUID
+
+	return log.Process("bootstrap", "Cloud infrastructure", func() error {
+		baseRunner := terraform.NewRunnerFromConfig(metaConfig, "base-infrastructure", stateCache).
+			WithVariables(metaConfig.MarshalConfig()).
+			WithAutoApprove(true)
+		tomb.RegisterOnShutdown("base-infrastructure", baseRunner.Stop)
+
+		_, err := terraform.ApplyPipeline(baseRunner, "Kubernetes cluster", terraform.GetBaseInfraResult)
+		return err
+	})
+}
